+++ packages/app/src/server/middlewares/api-rate-limiter.ts
@@ -3,9 +3,22 @@ import { RateLimiterMemory } from 'rate-limiter-flexible';
 
 import loggerFactory from '~/utils/logger';
 
-import { IApiRateLimitConfig } from '../interfaces/api-rate-limit-config';
+import { generateApiRateLimitConfig } from '../util/generateApiRateLimitConfig';
+
 
 const logger = loggerFactory('growi:middleware:api-rate-limit');
+
+const defaultMaxPoints = 100;
+const defaultConsumePoints = 10;
+const defaultDuration = 1;
+const opts = {
+  points: defaultMaxPoints, // set default value
+  duration: defaultDuration, // set default value
+};
+const rateLimiter = new RateLimiterMemory(opts);
+
+// generate ApiRateLimitConfig for api rate limiter
+const apiRateLimitConfig = generateApiRateLimitConfig();
 
 const consumePoints = async(rateLimiter: RateLimiterMemory, key: string, points: number, next: NextFunction) => {
   await rateLimiter.consume(key, points)
@@ -17,14 +30,14 @@ const consumePoints = async(rateLimiter: RateLimiterMemory, key: string, points:
     });
 };
 
-module.exports = (rateLimiter: RateLimiterMemory, defaultPoints: number, apiRateLimitConfig: IApiRateLimitConfig) => {
+module.exports = () => {
 
   return async(req: Request, res: Response, next: NextFunction) => {
 
     const endpoint = req.path;
     const key = req.ip + req.url;
 
-    let points = defaultPoints;
+    let points = defaultConsumePoints;
     Object.keys(apiRateLimitConfig).forEach((endpointInConfig) => {
       if (endpointInConfig === endpoint) {
         const consumePointsInConfig = apiRateLimitConfig[endpointInConfig].consumePoints;
